AVA/requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.4.2
requests==2.31.0

# Google Cloud & AI
google-api-core>=2.11.0
google-auth>=2.17.3
google-cloud-storage==2.12.0
google-cloud-aiplatform==1.36.4
google-cloud-vision==3.4.5
google-cloud-videointelligence==2.13.4
google-generativeai==0.3.2  # Gemini API
vertexai>=1.17.0  # VertexAI (for other models)

# Document processing
PyMuPDF==1.23.7  # For PDF processing
python-docx>=0.8.11  # DOCX processing
pdf2image==1.16.3  # Convert PDF to images
pytesseract==0.3.10  # OCR for text extraction
nltk==3.8.1  # Text analysis

# Image and video processing
Pillow==10.1.0  # Image processing
opencv-python==4.8.1.78  # Image and video processing
moviepy==1.0.3  # Video processing

# ML and Data processing
numpy==1.26.1
pandas>=2.0.0
torch==2.1.1
transformers==4.35.2